clarification for in-tree migration
Changes made to the lines 70 and 72.
current revision sounds a little contradictory when says that is not imperative to migrate in line 72, while in line 70 it said that "you will need to migrate", as if the PVCs would fail if you don't perform the migration.

After the migration, if the customer keeps the in-tree drivers, they will continue to work. But they will lose features.